+++ backend/frontend/src/components/navigation/Navbar.jsx
@@ -12,7 +12,7 @@ import { Login } from '../Login'
 
 function BasicsNavbar() {
   return (
-  <BrowserRouter> 
+  // <BrowserRouter> 
     <>
       <Navbar bg="success" variant="dark" expand="lg">
         <Container>
@@ -20,8 +20,8 @@ function BasicsNavbar() {
           <Navbar.Toggle aria-controls="basic-navbar-nav" />
           <Navbar.Collapse id="basic-navbar-nav" className="justify-content-end">
             <Nav className="ml-auto">
-              <Nav.Link as={Link} to='/Home'> Home </Nav.Link>
-              <Nav.Link as={Link} to='/Login'>User Profile</Nav.Link>
+              <Nav.Link as={Link} to='/home'> Home </Nav.Link>
+              <Nav.Link as={Link} to='/login'>Login</Nav.Link>
               <NavDropdown title="Settings" id="basic-nav-dropdown">
                 <NavDropdown.Item href="#Settings/Visuals">Visuals</NavDropdown.Item>
                 <NavDropdown.Item href="#Settings/Sound">
@@ -35,13 +35,13 @@ function BasicsNavbar() {
       </Navbar>
       
       <div>
-        <Routes>
+        {/* <Routes>
           <Route path='Login' element={<Login/>}></Route>
-        </Routes>
+        </Routes> */}
 
       </div>
     </>
-  </BrowserRouter>
+  // </BrowserRouter>
   );
 }
